DELETE POST

+++ controllers/postControllers.js
@@ -125,29 +125,36 @@ const editPost = async (req, res, next) => {
         } else {
             // get old post from database
             const oldPost = await Post.findById(postId);
-            // delete old thumbnail from upload
-            fs.unlink(path.join(__dirname, '..', 'uploads', oldPost.thumbnail), async (err) => {
-                if (err) {
-                    return next(new HttpError(err))
-                }
-            })
+            if (req.user.id == oldPost.creator) {
+                // if{!req.files}{
+                //     updatedPost = await Post.findByIdAndUpdate(postId, { title, category, description}, { new: true })
+                // } else{
 
-            // upload new thumbnail
-            const { thumbnail } = req.files;
-            // check file size
-            if (thumbnail.size > 2000000) {
-                return next(new HttpError("Thumbnail too big. Should be less than 2mb"))
+                // }
+                // delete old thumbnail from upload
+                fs.unlink(path.join(__dirname, '..', 'uploads', oldPost.thumbnail), async (err) => {
+                    if (err) {
+                        return next(new HttpError(err))
+                    }
+                })
+
+                // upload new thumbnail
+                const { thumbnail } = req.files;
+                // check file size
+                if (thumbnail.size > 2000000) {
+                    return next(new HttpError("Thumbnail too big. Should be less than 2mb"))
+                }
+                fileName = thumbnail.name;
+                let splittedFilename = fileName.split('.')
+                newFilename = splittedFilename[0] + uuid() + '.' + splittedFilename[splittedFilename.length - 1]
+                thumbnail.mv(path.join(__dirname, '..', 'uploads', newFilename), async (err) => {
+                    if (err) {
+                        return next(new HttpError(err))
+                    }
+                })
+
+                updatedPost = await Post.findByIdAndUpdate(postId, { title, category, description, thumbnail: newFilename }, { new: true })
             }
-            fileName = thumbnail.name;
-            let splittedFilename = fileName.split('.')
-            newFilename = splittedFilename[0] + uuid() + '.' + splittedFilename[splittedFilename.length - 1]
-            thumbnail.mv(path.join(__dirname, '..', 'uploads', newFilename), async (err) => {
-                if (err) {
-                    return next(new HttpError(err))
-                }
-            })
-
-            updatedPost = await Post.findByIdAndUpdate(postId, { title, category, description, thumbnail: newFilename }, { new: true })
         }
         if (!updatedPost) {
             return next(new HttpError("Couldn't update post.", 400))
@@ -159,44 +166,38 @@ const editPost = async (req, res, next) => {
 }
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 // ========================= DELETE POST
 // DELETE : api/posts/:id
 // PROTECTED
 const deletePost = async (req, res, next) => {
-    res.json("Delete Post")
+    try {
+        const postId = req.params.id;
+        if (!postId) {
+            return next(new HttpError("Post unavailable.", 400))
+        }
+        const post = await Post.findById(postId);
+        const fileName = post?.thumbnail;
+        if (req.user.id == post.creator) {
+            // delete thumbnail from uploads folder
+            fs.unlink(path.join(__dirname, '..', 'uploads', fileName), async (err) => {
+                if (err) {
+                    return next(new HttpError(err))
+                } else {
+                    await Post.findByIdAndDelete(postId);
+                    // find user and reduce post count by 1
+                    const currentUser = await User.findById(req.user.id);
+                    const userPostCount = currentUser?.posts - 1;
+                    await User.findByIdAndUpdate(req.user.id, { posts: userPostCount })
+                    res.json(`Post ${postId} deleted successfully.`)
+                }
+            })
+        } else {
+            return next(new HttpError("Post could not be deleted.", 403));
+        }
+    } catch (error) {
+        return next(new HttpError(error))
+    }
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 
 module.exports = { createPost, getPosts, getPost, getUserPosts, getCatPost, editPost, deletePost }
